Consume all events in a touchscreen gesture sequence
Attempt to address #26.

It is possible that some Android systems stop sending a sequence of
gestures as soon as one motion event is not consumed. This change means
all events in a touch screen sequence are now consumed, hopefully
meaning that we always get the ACTION_UP to finish the gesture.

Additionally, handled ACTION_CANCEL as an alternative to ACTION_UP.