Reduce cluetabs data invalidation
There was an artefact when changing from history to down, selecting a
clue would reset the view to the start. This is because it was creating
a brand new down view (even though it was already showing one).

Make each tab listen individually to the board (they cannot unlisten,
but should not cause memory leaks because a weak set is used). Across
and down can use ClueList to update only the clue views needed. History
list still needs to invalidate its whole list because it can't know the
old position of the newly selected clue without maintaining it's own
list. Didn't seem worth it.

Scroll positions seem to stay as they should now.