Correct snap to clue behaviour
Remove notion of max scroll because it only worked in one direction, was
only used by clue snapping, and caused clue snapping to behave
erratically.

Specifically, if a word was scrolled off the right (or bottom) it would
jump the clue as far left as possible, rather than just enough to see
the whole word on screen.

ensureVisible now simplified and seems to behave ok.