Fix date format in IPuzIO
Was using British rather than US ordering.

IPuzIO is now a bit bodged as all previously written ipuz files have the
date in the wrong order. Added a version filed to forkyz extensions so
that we can detect if it's the old one (i.e. it has no version field,
but does have playdata).