Fix toggle direction bug on entry
When stopping at the end of the word, the position is reset if the next
free space is outside the word. However, the search may also have
toggled the direction, so this needs resetting too.

This doesn't seem to come up in block puzzles as the last square in the
direction is usually part of a clue in the same direction. It seems more
common in bar puzzles that the last square is not part of a clue in the
same direction.