Detect cancelled gestures in ScrollingImageView
When using gesture navigation on the device, if the user, e.g. scrolls up
from the bottom to hide the app, the image view still gets the scroll
events up until the system cancels it. This meant the board would scroll
a little also when the user did not intend it.

Make ScrollingImageView watch out for the gesture cancel, and reset the
scroll location if it happens.