Remove LAT and WaPo Classic
Both have added a required "pickerToken" to prevent direct downloads.
The WaPo Sunday is still open.

The picker token could be obtained from the puzzle selection page (ala
xword-dl), but this feels a bit too much like breaking in for me.

It's still possible to download with xword-dl and import via the + menu.

Part of #40