Don't autonumber formats that have numbers
They may not always be "standard" crossword numbering, allow for
non-standard numbering.

There is still some technical debt: we cannot tell if a numbered cell is
the start of an across or down clue without the clue list. Currently
making "standard" assumptions to infer whether it is the start. Better
would be to set the boxes and clues and the same time so the data can be
filled in properly.