Move Position to Puzzle, use row/col, keep clues/boxes synced
Instead of assuming which numbers are attached to clues in a given
direction, fill in box data on whether the box is part of a clue from
existing clues, and as clues are added.

Doing this sensible meant moving the wordStarts maps from Playboard to
Puzzle, along with the Position class. And while doing it, switch
Position to use row/col instead of across/down for app consistency.