Convert background puzzle download to WorkManager
Warning: ungraceful migration from JobScheduler in that it will complain
about a job scheduled for a class that doesn't exist anymore. After that
i think it deletes it.

Will do better in a future commit when download period is a preference.
Plan to add a check to BrowseActivity to detect if old preference was
set, migrate it to WorkManager, and cancel any JobScheduler tasks.