Support split clues in Guardian
Made the decision to only include the full zone on the initial clue, and
not on the sub-clues. This means the sub-clues can't be directly
selected, but means dontDeleteCrossing doesn't think there are three
clues sharing the same zone.

Perhaps make dontDeleteCrossing smarter in future. But then this
behaviour matches how the Independent encode their split clues (i think).

Closes #49