feat(ui): add option to view board for all clues in clue list
At base -- add a BoardWordEditView to each clue list item in ClueTabs
that can optionally be shown. Add feature to show a specific word rather
than the current word in BoardWordEditView.

It is tricky to get keyboard input through the ClueTabs to the board
view, and also bad practice because the recycler view will reuse the
same board view for multiple clues. Instead, have a "hidden" board view
still at the top of the clue list that actually takes the focus and
input.