bug(ui): fix board freeze on clue list activity
Occurred because the "word positions" cache wasn't being cleared when
the current word changed (and the BoardWordEditView was rendering the
current word). It carried on using the old cache and only updating for
the original word.

For some reason required going to the notes page, then clicking off the
board, then returning to the clue list and changing clues... Not sure
why it didn't happen when changing any clue on the clue list activity.