feat(io): make winword own list/clue in RaetselZentraleSchweden IO
The old technique of adding extra cells to the bottom of the grid is a
bad user experience: the cells are not connected and don't mirror the
main grid.

Making it a clue means you can select the whole word, navigate from box
to box, and see the boxes together on the clues list and notes page.

Next step: implement a "focussed" word to display where the old extra
cells used to be.

Abandoned attempt to add a separate "solution zone" to the Puzzle class.
It was just a special case everywhere in the code, in its own special
"null" clue list. I was basically implementing the solution above as its
own type of clue that doesn't fit the existing code.

Part of #57.