bug(model): prevent crash when zones contain duplicates
In a rare case, a zone contains the same letter twice. findZoneDelta
hit a stackoverflow/endless recursion.

Prevent this by adding a variant of findZoneDelta that uses the zone
index rather than position.

This is not a great fix -- the user experience is that the next/prev
movement gets trapped in the loop and needs user intervention to move
onto the parts of the clue outside of the loop.

However, avoiding this behaviour means rewriting the model all over so
that the current selection is not the highlighted box / clueID, but also
the index of the highlighted box inside of the highlighted clue.

For now, a true fix is a bit hairy/time-consuming for what i think is a
very rare case.

At least it doesn't crash anymore :/