tweak(model): handle circular clues gracefully
When clues contain the same cell twice in their zone, prevent zone
movement getting stuck in a loop by remembering the index of the last
position found by a zone movement. If the index still matches the
highlight position the next time the zone movement is called, start
searching from that index.