feat(ui): change layout for "acrostic" mode
Limit board view to match proportions of actual board (may be long and
thin). Gives more room to clue tabs.

Words displayed on clue tabs. Because clue answers are spread around the
board, it is useful to have them together.

Set clue tabs to be at least 1/3 of the screen height (it takes space
away from the board view when keyboard is showing). Need to be able to
see both as you might be inputting an answer, or the full quote.

Need to adjust clue click notifications as behaviour has to change
slightly when board view clicked on clue tabs ("previous" word is not
the current word as the board click might have changed it already).